backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
# Token auth is built into DRF (rest_framework.authtoken)
# For Random Forest (optional; falls back to constant risk if not installed)
scikit-learn>=1.3
# numpy is auto-installed by scikit-learn
# networkx (optional) - for advanced graph algorithms if needed
# networkx>=3.0
# Testing
pytest>=7.0
pytest-django>=4.5